requirements.txt
# Core
streamlit==1.27.0
python-dotenv==1.0.0
hashlib

# LangChain
langchain==0.2.21
langchain-core==0.2.21
langchain-community==0.2.21
langchain-huggingface==0.0.19

# Document loaders
unstructured==0.9.7
pypdf==3.14.0

# Vector store
chromadb==0.3.28
sentence-transformers==2.2.2
torch==2.3.0
transformers==4.32.0

# Optional (improves DOCX parsing)
python-magic==0.4.27
